Move component builders directory
We use singular for package names
