+++ rest/src/main/kotlin/builder/component/ButtonBuilder.kt
@@ -1,4 +1,4 @@
-package dev.kord.rest.builder.components
+package dev.kord.rest.builder.component
 
 import dev.kord.common.annotation.KordDsl
 import dev.kord.common.annotation.KordPreview
