+++ rest/src/main/kotlin/builder/interaction/UpdateMessageInteractionResponseCreateBuilder.kt
@@ -2,13 +2,12 @@ package dev.kord.rest.builder.interaction
 
 import dev.kord.common.annotation.KordPreview
 import dev.kord.common.entity.InteractionResponseType
-import dev.kord.common.entity.MessageFlag
 import dev.kord.common.entity.MessageFlags
 import dev.kord.common.entity.optional.*
 import dev.kord.common.entity.optional.delegate.delegate
 import dev.kord.rest.builder.RequestBuilder
-import dev.kord.rest.builder.components.ActionRowBuilder
-import dev.kord.rest.builder.components.MessageComponentBuilder
+import dev.kord.rest.builder.component.ActionRowBuilder
+import dev.kord.rest.builder.component.MessageComponentBuilder
 import dev.kord.rest.builder.message.AllowedMentionsBuilder
 import dev.kord.rest.builder.message.EmbedBuilder
 import dev.kord.rest.json.request.InteractionApplicationCommandCallbackData
